refactor: :recycle: move delegate back into messaging example
