+++ plugins/flutter_aepmessaging/pubspec.yaml
@@ -14,10 +14,6 @@ dependencies:
     sdk: flutter
   flutter_aepcore: ">= 3.0.0 <4.0.0"
 
-dependency_overrides:
-  flutter_aepcore:
-    path: ../flutter_aepcore
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
